Parse links from the response header.
This changes the 'next' link parsing logic to parse links from the
response header instead of the response json. This makes it so that
any new links that Orchestrate eventually adds (ie 'prev') will
be automatically picked up by the client. Also, as pagination is
added to other parts of orchestrate (events, graph, etc), the client
also picks it up automatically.

Any links found in the 'Link' header on any 'GET' request will be
parsed out and made available on the response object under
response.links.  Each link will have various properties (url, rel,
params, etc) as well as a 'get' function to fetch the url that
returns a promise.